requirements-0.1.0.txt
click==8.0.3
jsbeautifier==1.15.1
lxml==5.3.0
napkinxc==0.6.2
numpy==1.24.3
pandas==2.2.2
pyreadr==0.5.2
pystempel==1.2.0
pytorch_lightning==1.9.5
sacremoses==0.1.1
scikit_learn==1.2.2
scipy==1.8.0
stop_words==2018.7.23
torchmetrics==1.4.1
tqdm==4.66.5
transformers==4.44.2
